formatted stat numbers

+++ src/main/kotlin/io/github/zabuzard/discordplays/Config.kt
@@ -20,7 +20,7 @@ data class Config(
     var bannedUsers: Set<Snowflake> = setOf(),
     var hosts: Set<HostId> = setOf(),
     var autoSaveRemindAt: LocalTime = LocalTime(13, 0),
-    var userToInputCount: List<Pair<UserSnapshot, Int>> = emptyList(),
+    var userToInputCount: List<Pair<UserSnapshot, Long>> = emptyList(),
     var playtimeMs: Long = 0,
     var recordingPath: String = "recording",
     var font: String = "Sans Serif"
